requirements.txt
numpy==2.4.3
pymongo==4.16.0
twilio==9.10.4
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
azure-functions==1.18.0